Interact
Okay so I created an Interact script that will allow the player to interact with the shop and other items if necessary. Whoever is going to add the interactable shop and other items, make sure these items are in the InteractabelLayer so the player can interact with it and make sure the names are all the same!