RoomManagement and EnemySpawner
In these scripts, they talk to each other but the jist of it is that the enemies will spawn as the player walks in and only three at a time, and 10 enemies in total, and 3 waves in total. Once the room in complete the door will unlock for the player to walk out! Who ever is in charge of the map you might need to make spawn points and attack the colliders to them from the scripts!